pyproject.toml
[project]
name = "capture-region-reader"
version = "0.1.0"
description = "Screen region OCR with real-time text-to-speech reader"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.7",
    "pytesseract>=0.3.13",
    "pyttsx3>=2.98",
    "mss>=9.0",
    "pynput>=1.7.7",
    "Pillow>=10.0",
    "numpy>=1.26",
]

[project.scripts]
capture-region-reader = "capture_region_reader.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/capture_region_reader"]
